more alpha channel support.
I will have to look at some point if there is any benefit from dealing with alpha write in a different path. Currently, for simplicity, all filters down to the WritePixel function. This means a lot of "if's". :/
But is easy to maintain. So look at it if I have to.